+++ components/web3-provider.js
@@ -1,4 +1,5 @@
 import { Global } from "@emotion/core";
+import WalletConnectWeb3Provider from "@walletconnect/web3-provider";
 import Authereum from "authereum";
 import {
   createContext,
@@ -32,6 +33,12 @@ export const createWeb3 = (infuraURL) => {
   web3.modal = new Web3Modal({
     cacheProvider: true,
     providerOptions: {
+      walletconnect: {
+        package: WalletConnectWeb3Provider,
+        options: {
+          infuraId: infuraURL.slice(infuraURL.lastIndexOf("/") + 1),
+        },
+      },
       authereum: {
         package: Authereum,
       },